// src/app/models/create-appointment.model.ts
export class CreateAppointmentModel{
    startDate: string = "";
    endDate: string = "";
    doctorId: string = ""; 
    patientId: string = "";
    firstName: string = "";
    lastName: string = "";
    identityNumber:string = "";
    city: string = "";
    town: string = "";
    fullAddress: string = ""; 
}